♻️ refactor : paymentService와 ReserveService 병합
